Remove EC_KEY_(get,set)_conv_form.
From codesearch, nothing uses this API.

Having an EC_KEY remember how it was encoded for reencoding purposes
seems somewhat pointless. Always encoding in uncompressed should be a
safe simplification. Nothing can be depending on keys re-encoding
completely identically as OpenSSL's legacy ASN.1 stack accepts more
encodings than it produces.

Having a stateful enc_flags on the EC_KEY is similarly unnecessary, but
varying EC_PKEY_NO_PARAMETERS is actually valuable. ECPrivateKey is
encoded standalone with parameters but typically omits it in PKCS#8
(where the parameters go in the algorithm). So that will wait until the
follow-up commit to rewrite ECPrivateKey serialization with a new API.

Change-Id: I43121070a7e3e4b00a4bdb5ac3412ad41245164c