Update patch set 5
Patch Set 5:

> This shouldn't have any practical consequences, except if a server
switched from not supporting tickets and then started supporting it,
while keeping the session cache fixed.

The client supporting session tickets but retaining only a session ID is a potential indication that the client's connection was downgraded from TLS 1.x to SSL 3.0 in the connection that established the session, or other similar badness. 

IMO, it would be useful to check that the resumed session's version is the same version as what you would negotiate without resumption, to reduce the potential of a successful downgrade attack. Maybe this check is made; I don't know, as I read only the patch, which doesn't mention this.

HTH.

Patch-set: 5
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0