Update patch set 2
Patch Set 2:

I think the situation with alignas might actually be more ridiculous than that. C++11 makes it a keyword, but C11 makes it a macro so you need both a C11-capable compiler and standard library.

I suspect one of our many toolchains will have a problem with this. +agl to toy with Android when he has time. I've uploaded https://codereview.chromium.org/1617243002 for Chromium to see.

Patch-set: 2