Update patch set 2
Patch Set 2:

> `OPENSSL_cleanse(NULL, 100)` (i.e. second parameter not zero) does
 > nothing in OpenSSL but causes problems in BoringSSL. It's not just
 > the `memset(NULL, _, 0)` problem. Internally in BoringSSL, it would
 > be bad to allow `OPENSSL_cleanse(NULL, x)` for `x != 0` because
 > that kind of behavior is indicative of a logic problem in the
 > caller that we'd want to fix. However, external code may be relying
 > on the OpenSSL behavior. That's why I split it into two separate
 > functions.

I don't think we've run into that one. Which external library does this?

It seems totally acceptable for OPENSSL_cleanse(NULL, 100) to crash, at least until we have some reason to believe it's a problem. That seems an incredibly stupid thing to call. If anyone's doing that, I think we'd not have a hard time convincing the upstream to stop. (Whereas OPENSSL_cleanse(NULL, 0) really should be okay.)

Patch-set: 2