Update patch set 2
Patch Set 2:

It seems bad to me to have a function with the same name as OpenSSL's function but which is less safe in any situation than the OpenSSL function. That is the basis for this change. But, if you don't agree with that line of reasoning then it's better to just abandon this.

Patch-set: 2
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0