Update patch set 2
Patch Set 2:

> It seems bad to me to have a function with the same name as
 > OpenSSL's function but which is less safe in any situation than the
 > OpenSSL function. That is the basis for this change. But, if you
 > don't agree with that line of reasoning then it's better to just
 > abandon this.

If it were a problem, we could just add an undocumented NULL check to the function. I think there's a couple of random accessors in ssl/ that we did that to. This one thus far hasn't been, and is pretty unlikely to.

OpenSSL historically put NULL checks obsessively, even when a function clearly shouldn't have one. It wouldn't make sense to make two versions of every API, just to remove a NULL check.

Patch-set: 2