Added protocol support for tlsext_maxfragment
This is the first of a two step process to add support for the TLS
maximum fragment length extension described in RFC 6066, section 4.
(https://tools.ietf.org/html/rfc6066#section-4).

This step adds the essential protocol support for the extension in the
ClientHello and ServerHello messages.  The fragment length is honored as
both sides will also set ssl->max_send_fragment as a result of
validating ssl->tlsext_max_fragment.

However, there's not yet any restriction on either the read_ or
write_buffer sizes.  The stated purpose of the TLS extension is to
enable constrained clients.  The second step will be to plumb the
ssl->tlsext_max_fragment value through to code sites currently using
SSL3_RT_MAX_PLAIN_LENGTH and similar.

Change-Id: I91849f676a8d3a0e78ab441fe186719c310e2f4c