Added protocol support for tlsext_maxfragment
This  adds support for the TLS maximum fragment length extension described
in RFC 6066, section 4. (https://tools.ietf.org/html/rfc6066#section-4).

It adds both the protocol support for the extension in the ClientHello and
ServerHello messages, as well as the library support by replacing the
various length definitions in ssl3.h with functions that derive the
proper values when a maximum fragment length has been negotiated.

The actual restriction on fragment size is achieved by simply using the
exisiting max_send_fragment field on both ends of the connection.

Change-Id: I91849f676a8d3a0e78ab441fe186719c310e2f4c