Remove REMAINDER_IS_ALREADY_CALCULATED logic from bn/div.c
That is defined iff div_asm is defined. It's simpler to say that div_asm is
always defined to touch rem and the non-div_asm path emulates both halves of
the instruction.

Also remove the #undefs. If we're worried that the environment might come with
div_asm defined, we should pick a different name since the later #ifdef will
break.

Change-Id: Ic2a64dc7c07588a2ac5e46e7c758477156e27b67