Update patch set 3
Abandoned

I'm abandoning this for now. It seems there's no way to use `STRICT_ALIGNMENT` without running afoul of the C strict aliasing rules. The endianness stuff should be redone in light of that and in light of David's comments, but it isn't urgent.

Thanks for the feedback, David!

Patch-set: 3
Status: abandoned