Update patch set 2
Abandoned

Abandoning this for now.

w.r.t. The constant-timedness of the RSA code, there are classes of issues: it isn't clear what is supposed to be constant-time, what needs to be constant -time, and what actually is constant-time. Also, although the rsaz code tries to avoid (cache) timing side channels, it isn't clear that it actually succeeds in doing so. Amd, also, the RSAZ code only works for a subset of the input value sizes for which side channel protection would be useful.

Thanks for the comments, Adam!

Patch-set: 2
Status: abandoned