Remove squaring optimizations for |BN_sqr| and |BN_mod_sqr|.
These functions are not called in performance-sensitive contexts, so it
doesn't make sense to carry the extra code for the optimized versions.

The biggest problem now is that these squaring functions are tested in
bn_test.cc by comparing their results to the results of using the of
the multiplication functions. But, now that's exactly how they are
defined, so the test is meaningless. Arguably, the test was almost
completely meaningless before, as these functions have always
(potentially) shared lots of code in their implementations.

Change-Id: Iaf9aae26627ef924315cd0a5100437f9ce39a69a