Update patch set 1
Patch Set 1:

Looking for suggestions on how to precede.

AFAICT, the main performance-sensitive function that uses these is `ec_GFp_simple_field_sqr`, which is used by p224-64.c and p256-64.c for the "is the point on the curve check." If that is a deal-breaker, then IMO the best solution is to accelerate the implementation of the custom "is the point on the curve check" function for those implementations using their special multiplication/squaring implementations.

I am also curious what you think about the issue with bn_test. What's the best way to resolve it?

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0