Update patch set 7
Abandoned

Abandoning. I think we can get rid of it and the other stateful one, but you can't actually implement PKCS#8 with the legacy APIs if we get rid of this. It's annoying, but not *that* in the way, so let's just leave it alone for now.

Patch-set: 7
Status: abandoned