Use stronger types for AES keys in AES-/mode-related code.
Use |const AES_KEY *| instead of |const void *| for the key type in
various type signatures. Originally the crypto/modes code was generic
over any 128-bit block cipher. Now it is only used for AES. Using the
stronger type reduces the amount of casting that needs to be analyzed
for correctness.

Change-Id: I1e9d2f65e38d0aac9c80a325debf6e995791d415