Update patch set 1
Patch Set 1:

Are you sure that you don't want to use OpenSSL in this case?

"Although BoringSSL is an open source project, it is not intended for general use, as OpenSSL is. We don’t recommend that third parties depend upon it. Doing so is likely to be frustrating because there are no guarantees of API or ABI stability."
(https://boringssl.googlesource.com/boringssl/)

We are fairly focused on meeting Google's needs with this code. If you're the maintainer of ZNC and want to start bundling BoringSSL in a hermetic build then this /might/ be reasonable. However, you might well be better off using the system OpenSSL shared library because that means that you can piggy-back on the normal update process.

(And, if you do move forward with this, we need contributors to sign the CLA: https://cla.developers.google.com/clas)

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0