Update patch set 8
Patch Set 8:

A separate EVP_PKEY_ type for PSS worries me.

For one, doing PSS signatures with traditional RSA SPKIs is going to be common. Also, RFC3447 includes different parameters for PSS and OAEP so would we end up with three PKEY types for RSA? Can't we, instead, make a note in the RSA* as to which SPKI type was used?

Patch-set: 8