Enable stitched AES-GCM asm optimizations on x86-64.
Enable AVX optimizations in the AES-NI assembly language code. Make the
stitching a hidden implementation detail of gcm.c instead of forcing
callers in e_aes.c to deal with it, so that all callers execute the
same logic. Previously, only the classic EVP interface for AES-GCM
had code for these optimizations while the AEAD interface didn't.

Note that the stitched implementations are no-ops for small inputs
(smaller than 288 bytes for encryption; smaller than 86 bytes for
decryption). This means that only a handleful of test cases with
longish inputs actually test the stitched code.

Change-Id: Iece8003d90448dcac9e0bde1f42ff102ebe1a1c9