Update patch set 1
Patch Set 1:

David, I'm not expecting this to be merged. I just shared it so you can use it to evaluate the AES-NI GCM stitching.

The numbers are all multiples of 96, which is what the stitching code is optimized for. That is, they show the best-case scenerio for the stitching code so that it is easier to evaluate whether it is worthwhile.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0