Update patch set 3
Patch Set 3:

My view is that it can always be changed back to `void *` later if it is necessary to do so.

My primary goal with this is to make it clear that there are no casting errors in BoringSSL, basically by removing all casting other than integer conversions. Part of that is demonstrating that there are no strict aliasing rule violations. But, another part of that is showing that there is no UB due to function casting. I think this kind of assurance is useful.

Besides that, my hypothesis is that there will be useful performance benefits to specializing the GCM code, in particular, for AES-not just the AESNI_GCM stitching, but other things. I've already got some of these improvements working and I think I'll have more working soon.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0