Update patch set 7
Patch Set 7:

I ran the BoringSSL tests on a Haswell machine before I uploaded the patch and they didn't catch this, for two reasons:

1. IIRC, there isn't a test of the EVP interface where, say, 289 (3 blocks + 1 byte, causing the 1 byte to be buffered until the next call) is encrypted and then later more than 15 bytes (<=15 bytes isn't enough to trigger it) are encrypted before set_iv is called.

2. More generally, there are not tests for the case where 288+(16*n)+m, m < 15 bytes are encrypted, followed more encryptions of various sizes without changing the IV (again only possible with the code in EVP interface).

3. The code would also become more incorrect if `GHASH_CHUNK` were changed to be less than 288/96 such that the `while (len >= GHASH_CHUNK)` loops iterate more than once.

The good news is that I think I just need to move the place where `ctr` is initialized like I did for the decrypt case; I will submit a patch that does at least this part.

It's also good news is that these sorts of problems occur only with the EVP interface (AFAICT). Since the EVP interface is exposed these conditions should be tested for it, but I won't have time to write those tests.

Regarding lengths: When I originally read the ASM code, I saw that $len was shifted right 4 bits and I noticed the return value is always a multiple of 16. But, I didn't notice how $end0 is used and I haven't had a chance to decipher the code to determine if it is safe. So, I think for now we should follow the example of the EVP code and mask off the last four bits before calling those functions. I will submit a patch that does this too.

Patch-set: 7