Clarify use of |$end0| in stitched AES-NI code.
There is a |cmp $in0,$end0| but there is no conditional jump
after the comparison. it turns out that |$end0| is now only used for
the key (cache?) aliasing check. This is important because it makes it
clearer that it is OK to pass values of |len| that aren't multiples of
16.

Change-Id: Iea546a59dc7aeb400f50ac5d2d7b9cb88ace9027