Split (almost) all public structs with CRYPTO_MUTEX or CRYPTO_refcount_t in two.
To avoid casting away _Atomic or having to guess on the size of
CRYPTO_MUTEX, split a number of structs into FOO and FOO_IMPL halves.
For compatibility, we cannot make all these structs opaque, so we'll
secretly always allocate a larger struct with trailing data. If we later
opaquify these types (notably, I hope to eventually opaquify SSL_CTX),
we can undo the split.

This should be legal even per a pedantic reading of the C specification.
C99 6.7.2.1 paragraph 13 says:

  A pointer to a structure object, suitably converted, points to its
  initial member (or if that member is a bit-field, then to the unit in
  which it resides), and vice versa. There may be unnamed padding within
  a structure object, but not at its beginning.

I interpret this to mean that C specifically allows one to cast pointers
in this way. (I'm unclear if the Linux kernel's more general
container_of trick is Officially legal rather than merely morally so,
but we don't need it.)

This converts everything except for some crypto/x509 types which
participate in crypto/asn1's more complex machinery. That will be
handled in the next change.

Note that this means these structs now may not be stack-allocated and
used with functions which touch locks or refcounts. Given that they
don't provide _init/_cleanup and only _new/_free, I cannot think of any
way this could have worked before. (Maybe if zero-initializing a
CRYPTO_MUTEX happened to work...)

BUG=590442

Change-Id: I4246b076702cf0d02b08338fcdca82062341b959