Update patch set 1
Patch Set 1: Code-Review-1

I think the conclusion the last time I spoke to Chandler about this is that it might be possible to do with lots of volatiles but that, currently, we don't have enough and that a sufficiently "smart" compiler could break this and several other of our constant-time functions.

I would tend to shy away from this change because I don't think it's well founded: the arguments are now volatile but the local-variable pointers are not. I think if we wanted to do this the path would be to make a dummy change in google3 that adds volatile all over this function, send it to Chandler and see which he thinks could be removed (if any).

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1