Update patch set 5
Patch Set 5:

> > I verified via bisect that this broke x86. See https://boringssl-review.googlesource.com/#/c/7198/.
 > 
 > Lovely. :-/ Thanks! I'll look into this.

If you're curious, Brian, it looks like this is because upstream writes blocks out in order 0, 4, 8, 9, 12, 14, 1, 2, 3, 5, 6, 7, 10, 11, 14, 15. (In the non-SSE3 code. I haven't looked at what's up with the SSE3 code, but I'm guessing it's similar. non-SSE3 repros as soon as we have a whole block of input. SSE3 seems to require much larger, like 256 or so.)

Patch-set: 5