Update patch set 2
Patch Set 2:

If I remove this code then things do break for me. I don't know why you didn't see it, but this does appear to be meaningful.

Thus the question remains whether the length can be a non-multiple of 16. The code if obviously working on multiples of 16 so I don't believe that it could ever process a partial block. Thus there doesn't appear to be anything lost if we mask out ~15 when passing in the length. I'll do a change to make that happen.

Patch-set: 2