Only pass a multiple of 16 to aesni_gcm_[en|de]crypt.
The assembly code in aesni-gcm-x86_64.pl operates on blocks of 16 bytes
and never processes a partial block. However, it calculates an end
pointer by adding the length to the start. This makes it unclear whether
passing in a length that is not a multiple of 16 is safe and so this
change ensures that we don't do that.

Change-Id: I5d772a27ad1056e41e7ceac8110d08b2fc5d45c3