Update patch set 1
Patch Set 1: Code-Review-1

If you look at the current OpenSSL code, you can see that it doesn't attempt to avoid passing partial blocks to these functions. So, it is strange that BoringSSL attempts to do so.

Also, despite your comment in the other bug, when I run the current (up to date as of this moment) test suite, all the tests pass, even with my "remove all the confusing logic" patch. So, that means that we should add some new test cases that show why it is important to make this change.

It is better to understand and document what the confusing code is really doing. I now understand it. it is actually very simple. I'll submit a changeset that adds a comment describing it.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1