Update patch set 3
Patch Set 3:

The tests won't fail when we tweak these things because this only prevents a read buffer overflow, which doesn't affect the calculation of the result unless some kind of fault occurs due to the out-of-bounds read. In order to test for this, we'd need to add infrastructure that lets us detect read buffer overflows that we don't currently have.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0