Add TLS transcripts from our current tests to the fuzz corpus.
This isn't perfect as the fuzzer won't have session caching or random features
enabled, but it should be something. (Probably the right thing to do is to
enable almost features in the fuzzer driver. Or perhaps we start the input with
some feature bits if we think there's interesting codepaths on the disabling
side.)

We don't currently have DTLS fuzzers, so I've left the DTLS transcripts alone.
We can copy them over when we have them.

Change-Id: Ie54fd442df2c047e838027a0df45fb5f6e79b5bc